[ESSI-588] [ESSI-446] Refactor config/initializers using prepend
