+++ config/application.rb
@@ -13,7 +13,8 @@ module ESSI
     # Initialize configuration defaults for originally generated Rails version.
     config.load_defaults 5.1
 
-    config.autoload_paths += ['app/models/vocab']
+    config.autoload_paths += ['app/models/vocab', Rails.root.join('lib')]
+    config.eager_load_paths += ['app/models/vocab', Rails.root.join('lib')]
 
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration should go into files in config/initializers
