--- config/initializers/add_file_manager_form_fields.rb
@@ -1,7 +1,0 @@
-module FileManagerFormExtensions
-  delegate :viewing_direction, :viewing_hint, to: :model
-end
-
-Hyrax::Forms::FileManagerForm.class_eval do
-  prepend FileManagerFormExtensions
-end
