+++ lib/extensions/hyrax/forms/file_manager_form/viewing_metadata.rb
@@ -1,0 +1,11 @@
+module Extensions
+  module Hyrax
+    module Forms
+      module FileManagerForm
+        module ViewingMetadata
+          delegate :viewing_direction, :viewing_hint, to: :model
+        end
+      end
+    end
+  end 
+end
